refactor: widen repo remove id to u64, include filename in duplicate id error, filter git list by pinned ref, and extract shared test setup helper

+++ crates/core-node-internal/schemas/repo.capnp
@@ -97,7 +97,7 @@ struct RepoListResponse {
 
 struct RepoRemoveRequest {
     # ID of the repository to remove
-    id @0 :UInt32;
+    id @0 :UInt64;
 }
 
 struct RepoRemoveResponse {
